Moves reading of settings into config file
